Teams uses keys internally and names externally
Names are nice and pretty, they have spaces and will consistently use
title case. Keys are useful, they are downcased and can be stuck in a
url. We only store one value, to avoid the confusion that might arrise
when they get out of sync. This does mean that the user will find that
'My FAvOURiTe pipleine' becomes 'My Favourite Pipleine', but this is
probably better than the confusion cause when 'My Pipeline!' is treated
the same as 'My Pipline.' The other option would be to disambiguate
the keys ourselves with addition of numbers, but having just one
name for stuff is probably best over all.

And yes, that was just a whole paragraph on the ux of names and
resource identifiers. Look forward to the completely contradictory
commit message later on in the project when I change my mind.