Generate almost safe keys from names
Keys will ge generated automatically from names, and will be counter
incramented in the event of a clash. This solution still leaves us
open to potential race conditions, which can be handled by rescuing
ActiveRecord::RecordNotUnique. This is should be exceedingly rare
however and the database is protecting overall integrity.